python/insight/pyproject.toml
[project]
name = "insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "gunicorn==22.0.0",
    "influxdb==5.3.2",
    "langchain-mcp-adapters==0.1.7",
    "langchain-ollama>=0.3.8",
    "langchain-openai==0.3.26",
    "langgraph==0.6.4",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-checkpoint-sqlite",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.9",
    "langmem==0.0.29",
    "mcp==1.9.4",
    "numpy==1.26.4",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.34.1",
    "opentelemetry-exporter-otlp-proto-http>=1.34.1",
    "opentelemetry-instrumentation-fastapi==0.55b1",
    "opentelemetry-instrumentation-logging==0.55b1",
    "opentelemetry-instrumentation-requests==0.55b1",
    "opentelemetry-instrumentation-sqlalchemy==0.55b1",
    "pandas==2.3.0",
    "prophet==1.1.5",
    "pymysql==1.1.1",
    "pytz==2024.1",
    "rrcf==0.4.4",
    "setuptools==72.1.0",
    "sqlalchemy==2.0.34",
    "uvicorn==0.30.6",
]

[tool.uv.workspace]
members = [
    "external/checkpoint-sqlite",
]

[tool.uv.sources]
langgraph-checkpoint-sqlite = { workspace = true }

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.2",
]

[tool.ruff]
# Line length setting (Black default)
line-length = 120
indent-width = 4

# Python version
target-version = "py312"

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "external/",
    "checkpoints/",
]

[tool.ruff.lint]
# Enabled rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort (import sorting)
    "B",      # flake8-bugbear
    "UP",     # pyupgrade (latest syntax by Python version)
    "C90",    # mccabe (complexity check)
    "N",      # pep8-naming
    "RUF",    # Ruff-specific rules
]

# Ignored rules
ignore = [
    "E501",   # Line too long (controlled by line-length)
    "B008",   # Function call in argument defaults
    "C901",   # Too complex (adjust per function if needed)
    "N802",   # Function name should be lowercase
    "N803",   # Argument name should be lowercase
    "N806",   # Variable in function should be lowercase
    "N805",   # First argument of a method should be named self
    "N815",   # Variable in class scope should not be mixedCase
]

# Auto-fixable rules
fixable = ["ALL"]
unfixable = []

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # unused import, import *
"tests/**/*.py" = ["S101"]         # Allow assert usage
"**/models.py" = ["N805"]          # Allow pydantic/sqlalchemy model names

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.isort]
known-first-party = ["app", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# Quote style
quote-style = "double"

# Indentation
indent-style = "space"

# Line ending style
line-ending = "auto"

# Docstring code formatting
docstring-code-format = true
